# java-restserver/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample for swagger-maven-plugin rest",
    "version" : "v1",
    "title" : "Swagger Maven Plugin API REST"
  },
...
  "paths" : {
...
    "/rooms/{roomNumber}" : {
      "get" : {
        "tags" : [ "rooms" ],
        "operationId" : "getRoom",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "roomNumber",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RoomHotel"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "rooms" ],
        "operationId" : "bookRoom",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "roomNumber",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "RoomHotel" : {
      "type" : "object",
      "properties" : {
        "roomNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "BASIQUE", "MIDL_GAMME", "HAUT_GAMME", "UNRECOGNIZED" ]
        },
        "price" : {
          "type" : "number",
          "format" : "float"
        }
      }
    },
...
  }
}